WebVoyager/requirements.txt
openai==1.20.0
selenium
pyautogen==0.2.19
pillow
httpx==0.27.2
python-dotenv
sentence-transformers
chromadb
markdown-it-py